Add option to track multiple pages in one stream
