+++ src/main/scala/FacebookDemo.scala
@@ -6,7 +6,7 @@ object FacebookDemo {
     val mode = args.headOption.getOrElse("")
 
     // configure page for which to ingest posts
-    val pageId = "aljazeera"
+    val pageIds = List("aljazeera", "TheEconomist")
 
     // configure interaction with facebook api
     val auth = FacebookAuth(accessToken = System.getenv("FACEBOOK_AUTH_TOKEN"), appId = System.getenv("FACEBOOK_APP_ID"), appSecret = System.getenv("FACEBOOK_APP_SECRET"))
@@ -16,7 +16,7 @@ object FacebookDemo {
     Logger.getRootLogger.setLevel(Level.ERROR)
     Logger.getLogger("libfacebook").setLevel(Level.DEBUG)
 
-    if (mode.contains("standalone")) new FacebookDemoStandalone(pageId, auth).run()
-    if (mode.contains("spark")) new FacebookDemoSpark(pageId, auth).run()
+    if (mode.contains("standalone")) new FacebookDemoStandalone(pageIds.head, auth).run()
+    if (mode.contains("spark")) new FacebookDemoSpark(pageIds.toSet, auth).run()
   }
 }
