+++ src/main/scala/FacebookDemoSpark.scala
@@ -2,14 +2,14 @@ import com.github.catalystcode.fortis.spark.streaming.facebook.{FacebookAuth, Fa
 import org.apache.spark.streaming.{Seconds, StreamingContext}
 import org.apache.spark.{SparkConf, SparkContext}
 
-class FacebookDemoSpark(pageId: String, auth: FacebookAuth) {
+class FacebookDemoSpark(pageIds: Set[String], auth: FacebookAuth) {
   def run(): Unit = {
     // set up the spark context and streams
     val conf = new SparkConf().setAppName("Facebook Spark Streaming Demo Application")
     val sc = new SparkContext(conf)
     val ssc = new StreamingContext(sc, Seconds(1))
 
-    FacebookUtils.createPageStream(ssc, auth, pageId).map(x => s"Post: ${x.post.getPermalinkUrl}").print()
+    FacebookUtils.createPageStreams(ssc, auth, pageIds).map(x => s"Post: ${x.post.getPermalinkUrl}").print()
 
     // run forever
     ssc.start()
