+++ src/main/scala/com/github/catalystcode/fortis/spark/streaming/facebook/dto/FacebookPost.scala
@@ -3,6 +3,7 @@ package com.github.catalystcode.fortis.spark.streaming.facebook.dto
 import facebook4j.{Comment, Post}
 
 case class FacebookPost(
+  pageId: String,
   post: Post,
   comments: Seq[Comment]
 )
